refactor(style-engine): extract resolveMediaSettings into shared module
Move breakpoint min/max resolution out of use-media so header-ui iframe
preview sizing can reuse the same bounds as MediaQuery generation.
